requirements.txt
praw
configparser
langchain
langchain-community
langchain-core
langchain-openai
psycopg2-binary
